+++ database/migrations/2026_03_31_204055_create_member_roles_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
         Schema::create('member_roles', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('sort_order')->default(0);
+            $table->unsignedBigInteger('sort_order')->default(0);
             $table->boolean('is_active')->default(true);
             $table->timestamps();
         });
